Extract lifeEvents
